Safe delete folder.

+++ manteinance/delete-user-module.js
@@ -24,9 +24,11 @@ function runDelete(ids, callback) {
                     for (var i = 0; i < users.length; i++) {
                         var user = users[i];
 
-                        var realPath = path.join(config.steviaDir, config.usersPath, user.email);
+                        var realPath = path.join(config.steviaDir, config.usersPath, user.name);
                         try {
-                            shell.rm('-rf', realPath);
+                            if (shell.test('-e', realPath)) {
+                                shell.rm('-rf', realPath);
+                            }
                         } catch (e) {
                             console.log("File fsDelete: file not exists on file system")
                         }
